fix(mqtt,telnet): close the gaps a self-review found in beta.4 (TASK-1155)
A critical read of the TASK-1148 and TASK-1154 work after beta.4 shipped found
five defects. All five are fixed here.

1. mqtt_desync_drops counted 2 of 14 desync sites. The PROGMEM payload site,
   the sendMQTT() payload site and all ten discovery-composer sites dropped the
   link without counting, and the composers send the largest payloads, so the
   counter read low for exactly the wrong reason. mrfox7688 had been asked to
   read it on GH #682. Every site now goes through mqttDropLinkOnDesync(), and a
   grep finds no bare desync disconnect left. The two remaining direct
   disconnects are the reboot path and a broker reconfiguration, neither a
   desync.

   The discovery composers also bypassed the send-buffer pre-flight entirely:
   beginDiscoveryPublish() called beginPublish() directly. It now pre-flights
   like every other path. A deferral is safe there because the discovery drip
   leaves the pending bit set on a false return and retries next tick; on the
   bench every deferred config was published within 15 s, disc_pending_ids 0.

   mqttDropLinkOnDesync() lives in mqtt_configuratie.cpp, not beside the other
   helpers, because the Arduino preprocessor hoists every .ino prototype above
   the includes, so a .ino signature cannot name PubSubClient. The counter bump
   stays in the .ino as mqttCountDesyncDrop(), which names no class.

2. The SimpleTelnet write retry could never run and is removed in the submodule
   (2a40633). Behaviour-neutral: a normally reading client dropped 0 bytes over
   25 s on the bench.

3. writeMqttChunk() and writeMqttProgmemChunk() retried up to ten times with no
   wall-clock bound, and each write can block for WiFiClient's 5000 ms
   no-progress timeout, so a stalled socket could hold the loop for about 50 s
   with the PIC serial line undrained. Both now stop after one such timeout
   (MQTT_WRITE_STALL_BUDGET_MS). Measured against a broker that stops reading,
   polling REST every second: longest unresponsive gap 4.9 s, once.

4. CLAUDE.md named the wrong re-entrancy path. feedWatchDog() does not yield,
   its yield() is commented out. Re-entrancy is nonetheless real, via delayms(),
   which spins doBackgroundTasks() and is reached from the telnet debug key b
   while already inside doBackgroundTasks(). The mechanism is corrected and the
   warning kept; removing it would have deleted a true warning.

5. iMqttSndbufSkips and iMqttDesyncDrops are now uint32_t, per the TASK-1039
   rule in the same struct that field-read counters must not wrap.

Reading the counters changes with this: a healthy boot now shows a few dozen
sndbuf deferrals, about 59 on the bench, because discovery is included. A stall
is the count still climbing after boot has settled.

Gates: build.bat green, fresh bins at 07:39 carrying 1.7.6-beta.5,
"Build completed successfully!", 0 errors. evaluate.py --quick 36/38, 0
failed. Host tests 76 checks across 4 suites, 0 failures. Static RAM
52732 -> 52740.

Refs #682

+++ src/OTGW-firmware/OTGW-firmware.h
@@ -348,8 +348,10 @@ struct HeapDiagSection {                 // state.heapdiag — per-boot heap-pre
   // consecutive seconds (WiFiClient _timeout = 5000, and ClientContext::_is_timeout()
   // resets its clock on every byte of progress), so both of these count severe
   // events, not routine backpressure.
-  uint16_t iMqttSndbufSkips         = 0; // publishes deferred because the header+topic would not fit in tcp_sndbuf
-  uint16_t iMqttDesyncDrops         = 0; // times the link was dropped to stop a half-written PUBLISH desynchronising the broker
+  // uint32 (TASK-1155): these are read over a churn period of days, and the TASK-1039
+  // note below applies to them too, a counter meant for field reading must not wrap.
+  uint32_t iMqttSndbufSkips         = 0; // publishes deferred because the header+topic would not fit in tcp_sndbuf
+  uint32_t iMqttDesyncDrops         = 0; // times the link was dropped to stop a half-written PUBLISH desynchronising the broker
   uint16_t iWsMaxBlockSkips         = 0; // WebSocket sends skipped by the maxBlock pre-flight gate (fragmentation guard)
   // TASK-1039: these two are LOOP-TICK / EVENT counters, not request counts, and they must
   // not wrap: field data shows ~365 ticks/s, so a uint16 wraps in roughly three minutes.
